+++ Sources/MapboxCoreNavigation/RouteController.swift
@@ -432,31 +432,7 @@ open class RouteController: NSObject {
         updateRouteLeg(to: routeProgress.legIndex + 1)
     }
     
-    /**
-     Path to the directory where history could be stored when `RouteController.writeHistory(completionHandler:)` is called.
-     */
-    public static var historyDirectoryURL: URL? = nil {
-        didSet {
-            Navigator.historyDirectoryURL = historyDirectoryURL
-        }
-    }
-    
-    /**
-     A closure to be called when history writing ends.
-     
-     - parameter historyFileURL: A path to file, where history was written to.
-     */
-    public typealias WriteHistoryCompletionHandler = (_ historyFileURL: URL?) -> Void
-    
-    /**
-     Store history to the directory stored in `RouteController.historyDirectoryURL` and asynchronously run a callback
-     when writing finishes.
-     
-     - parameter completionHandler: A block object to be executed when history writing ends.
-     */
-    public static func writeHistory(completionHandler: @escaping WriteHistoryCompletionHandler) {
-        Navigator.shared.writeHistory(completionHandler: completionHandler)
-    }
+    // MARK: Accessing Relevant Routing Data
     
     /**
      A custom configuration for electronic horizon observations.
@@ -485,6 +461,47 @@ open class RouteController: NSObject {
     /// The road object matcher that allows to match user-defined road objects.
     public var roadObjectMatcher: RoadObjectMatcher {
         return Navigator.shared.roadObjectMatcher
+    }
+    
+    // MARK: Recording History to Diagnose Problems
+    
+    /**
+     Path to the directory where history could be stored when `RouteController.writeHistory(completionHandler:)` is called.
+     */
+    public static var historyDirectoryURL: URL? = nil {
+        didSet {
+            Navigator.historyDirectoryURL = historyDirectoryURL
+        }
+    }
+    
+    /**
+     Starts recording history for debugging purposes.
+     
+     - postcondition: Use the `stopRecordingHistory(writingFileWith:)` method to stop recording history and write the recorded history to a file.
+     */
+    public static func startRecordingHistory() {
+        Navigator.shared.startRecordingHistory()
+    }
+    
+    /**
+     A closure to be called when history writing ends.
+     
+     - parameter historyFileURL: A URL to the file that contains history data. This argument is `nil` if no history data has been written because history recording has not yet begun. Use the `startRecordingHistory()` method to begin recording before attempting to write a history file.
+     */
+    public typealias HistoryFileWritingCompletionHandler = (_ historyFileURL: URL?) -> Void
+    
+    /**
+     Stops recording history, asynchronously writing any recorded history to a file.
+     
+     Upon completion, the completion handler is called with the URL to a file in the directory specified by `RouteController.historyDirectoryURL`. The file contains details about the route controller’s activity that may be useful to include when reporting an issue to Mapbox.
+     
+     - precondition: Use the `startRecordingHistory()` method to begin recording history. If the `startRecordingHistory()` method has not been called, this method has no effect.
+     - postcondition: To write history incrementally without an interruption in history recording, use the `startRecordingHistory()` method immediately after this method. If you use the `startRecordingHistory()` method inside the completion handler of this method, history recording will be paused while the file is being prepared.
+     
+     - parameter completionHandler: A closure to be executed when the history file is ready.
+     */
+    public static func stopRecordingHistory(writingFileWith completionHandler: @escaping HistoryFileWritingCompletionHandler) {
+        Navigator.shared.stopRecordingHistory(writingFileWith: completionHandler)
     }
 }