+++ Sources/MapboxCoreNavigation/SessionState.swift
@@ -16,8 +16,8 @@ struct SessionState {
     var numberOfReroutes = 0
     var lastRerouteDate: Date?
     
-    var currentRoute: Route
-    var originalRoute: Route
+    var currentRoute: Route?
+    var originalRoute: Route?
     
     var terminated = false
     
@@ -37,7 +37,7 @@ struct SessionState {
     
     var pastLocations = FixedLengthQueue<CLLocation>(length: 40)
     
-    init(currentRoute: Route, originalRoute: Route) {
+    init(currentRoute: Route? = nil, originalRoute: Route? = nil) {
         self.currentRoute = currentRoute
         self.originalRoute = originalRoute
     }