+++ Tests/MapboxCoreNavigationTests/PassiveLocationManagerTests.swift
@@ -34,7 +34,7 @@ extension CLLocationCoordinate2D {
     }
 }
 
-class PassiveLocationManagerTests: XCTestCase {
+class PassiveLocationManagerTests: TestCase {
     class Delegate: PassiveLocationManagerDelegate {
         let road: Road
         let locationUpdateExpectation: XCTestExpectation
@@ -64,8 +64,6 @@ class PassiveLocationManagerTests: XCTestCase {
     override func setUpWithError() throws {
         try super.setUpWithError()
         
-        Navigator.credentials = directions.credentials
-
         let bundle = Bundle(for: Fixture.self)
         let filePathURL: URL = URL(fileURLWithPath: bundle.bundlePath.appending("/tiles/liechtenstein"))
         Navigator.tilesURL = filePathURL
@@ -104,10 +102,11 @@ class PassiveLocationManagerTests: XCTestCase {
     func testNoHistoryRecording() {
         PassiveLocationManager.historyDirectoryURL = nil
         Navigator._recreateNavigator()
+        PassiveLocationManager.startRecordingHistory()
                 
         let noHistoryExpectation = XCTestExpectation(description: "History should not be written on 'nil' path")
         noHistoryExpectation.isInverted = true
-        PassiveLocationManager.writeHistory { _ in
+        PassiveLocationManager.stopRecordingHistory { _ in
             noHistoryExpectation.fulfill()
         }
         wait(for: [noHistoryExpectation], timeout: 3)
@@ -118,9 +117,10 @@ class PassiveLocationManagerTests: XCTestCase {
         
         PassiveLocationManager.historyDirectoryURL = supportDir
         Navigator._recreateNavigator()
+        PassiveLocationManager.startRecordingHistory()
                 
         let historyExpectation = XCTestExpectation(description: "History should be written to '\(supportDir)'")
-        PassiveLocationManager.writeHistory { url in
+        PassiveLocationManager.stopRecordingHistory { url in
             XCTAssertNotNil(url)
             historyExpectation.fulfill()
         }