+++ Tests/MapboxNavigationTests/CPMapTemplateTests.swift
@@ -1,9 +1,10 @@
 import XCTest
+import TestHelper
 @testable import MapboxNavigation
 import CarPlay
 
 @available(iOS 12.0, *)
-class CPMapTemplateTests: XCTestCase {
+class CPMapTemplateTests: TestCase {
     func testInitializationFromPanDirection() {
         XCTAssertEqual(CLLocationDirection(panDirection: .up), 0)
         XCTAssertEqual(CLLocationDirection(panDirection: .right), 90)